export/code/requirements.txt
Pillow
tensorflow==2.4.1